Add explicit cast in half_btf()
To silence asan failures in fuzzing tests.

BUG=:68825590,68825594,68825599

(cherry picked from commit e43ea91055133baaf3b691170a097a456c032e23)

+++ av1/common/av1_txfm.h
@@ -104,7 +104,7 @@ static INLINE void round_shift_array(int32_t *arr, int size, int bit) {
 
 static INLINE int32_t half_btf(int32_t w0, int32_t in0, int32_t w1, int32_t in1,
                                int bit) {
-  int32_t result_32 = w0 * in0 + w1 * in1;
+  int32_t result_32 = (int32_t)((int64_t)w0 * in0 + w1 * in1);
   return round_shift(result_32, bit);
 }
